requirements.txt
windows_curses==2.4.0; sys_platform == 'win32' or sys_platform == 'win64'
pycryptodome==3.21.0
pygame==2.6.1
scapy==2.6.0